Removed Trolly From Prefab
I removed the trolly function from virtual camera 1 in the prefab. Then I created an example folder that will have scenes that demonstrate what you can accomplish with cinnemachine. The goal is to have the base prefab simple and not have anything special so users can customize it as they see fit for their world.